+++ SmartPost.WPF/SmartPost.Domain/Entities/StokProducts/StokProduct.cs
@@ -2,11 +2,6 @@
 using SmartPost.Domain.Entities.Brands;
 using SmartPost.Domain.Entities.Categories;
 using SmartPost.Domain.Entities.Users;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SmartPost.Domain.Entities.StokProducts
 {
@@ -23,9 +18,9 @@ namespace SmartPost.Domain.Entities.StokProducts
 
         public string ProductName { get; set; }
         public string BarCode { get; set; }
-        public int PCode { get; set; }
-        public double Price { get; set; }
-        public int Quantity { get; set; }
+        public long PCode { get; set; }
+        public decimal Price { get; set; }
+        public decimal Quantity { get; set; }
         public string Status { get; set; }
     }
 }
