+++ SmartPost.WPF/SmartPost.Domain/Entities/StorageProducts/StorageProduct.cs
@@ -1,11 +1,6 @@
 ﻿using SmartPost.Domain.Commons;
 using SmartPost.Domain.Entities.Brands;
 using SmartPost.Domain.Entities.Categories;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SmartPost.Domain.Entities.StorageProducts
 {
@@ -19,8 +14,8 @@ namespace SmartPost.Domain.Entities.StorageProducts
 
         public string ProductName { get; set; }
         public string BarCode { get; set; }
-        public int PCode { get; set; }
-        public double Price { get; set; }
-        public int Quantity { get; set; }
+        public long PCode { get; set; }
+        public decimal Price { get; set; }
+        public decimal Quantity { get; set; }
     }
 }
